comp_scrapers/requirements.txt
beautifulsoup4==4.12.3
Scrapy==2.11.1
selenium==4.18.1
pandas==2.2.1
python-dotenv==1.0.1
SQLAlchemy==2.0.28
psycopg2-binary==2.9.9